Added some LibraryPath Utilities.
What it does:
The idea was to create something, which helps to "eliminate" .sh's and
.bat's.
How it should be done:
The Process, which got called via double-clicking (which is actually a
"java -jar javafile.jar", will look, whether the needed library path is
included in the java.library.path's. If it isn't, then it recreates a
JVM, calling itself with the right library path. ("java
-Djava.library.path="librarypath/" -jar javafile.jar")